renamed package
changed package name for entity from 'model' to 'entity'
